requirements.txt
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.60
chromadb==0.6.3
openai==1.59.8
pypdf==5.1.0
streamlit==1.41.1
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2